Optionally allow to reuse existing documents during reindexing
Change I1fd78958f6fa0848f81630abab0428520f412baa introduced a feature
to skip indexing of changes which were already present in the index.
This prevented unnecessary work, if Gerrit had to be shutdown while
performing an online reindex. However, this change also broke offline
reindexing, since this forcefully deleted all documents in the index
before reindexing, which wasn't transparent to the indexers.

This change is a reimplementation of the feature allowing to reuse
existing up-to-date documents in the index. The
`index.reuseExistingDocuments` option has to be set in the
`gerrit.config` to enable this feature. This works for offline and
online reindexing.

Existing but stale documents will be reindexed regardless.

The issue with offline reindexing is fixed, by not emptying the index
before indexing, if `index.reuseExistingDocuments` has been set.

Release-Notes: Optionally allow to reuse existing documents during reindexing

+++ java/com/google/gerrit/pgm/Reindex.java
@@ -258,7 +258,9 @@ public class Reindex extends SiteProgram {
     requireNonNull(
         index, () -> String.format("no active search index configured for %s", def.getName()));
     index.markReady(false);
-    index.deleteAll();
+    if (!globalConfig.getBoolean("index", null, "reuseExistingDocuments", false)) {
+      index.deleteAll();
+    }
 
     SiteIndexer<K, V, I> siteIndexer = def.getSiteIndexer();
     siteIndexer.setProgressOut(System.err);
